requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
amqp==5.3.1
annotated-types==0.7.0
anyio==4.11.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.10.0
async-timeout==5.0.1
attrs==25.4.0
autobahn==24.4.2
Automat==25.4.16
azure-cognitiveservices-speech==1.41.1
billiard==4.2.1
boto3==1.37.13
botocore==1.37.13
CacheControl==0.14.4
cachetools==6.2.2
catalogue==2.0.10
celery==5.5.3
certifi==2025.10.5
cffi==2.0.0
channels==4.3.1
channels_redis==4.3.0
charset-normalizer==3.4.4
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cloudpathlib==0.23.0
Collectfast==2.2.0
confection==0.1.5
constantly==23.10.4
crispy-bootstrap5==0.6
cron_descriptor==2.0.6
cryptography==44.0.0
cymem==2.0.11
daphne==3.0.2
defusedxml==0.7.1
dj-rest-auth==2.2.4
Django==5.2.7
django-allauth==0.51.0
django-anymail==13.1
django-celery-beat==2.8.1
django-ckeditor==6.5.1
django-constance==4.3.2
django-cors-headers==3.13.0
django-crispy-forms==1.14.0
django-environ==0.9.0
django-extensions==3.1.5
django-filter==25.2
django-js-asset==3.1.2
django-model-utils==4.2.0
django-nested-admin==4.1.1
django-recaptcha==3.0.0
django-redis==5.2.0
django-storages==1.14.6
django-timezone-field==7.1
django_celery_results==2.6.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.28.0
et_xmlfile==2.0.0
exceptiongroup==1.3.0
firebase_admin==7.1.0
frozenlist==1.8.0
future==1.0.0
google-api-core==2.28.1
google-auth==2.43.0
google-cloud-core==2.5.0
google-cloud-firestore==2.21.0
google-cloud-storage==3.6.0
google-crc32c==1.7.1
google-resumable-media==2.8.0
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.76.0
gunicorn==22.0.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
hyperlink==21.0.0
idna==3.11
incremental==24.7.2
inflection==0.5.1
Jinja2==3.1.6
jmespath==1.0.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kombu==5.5.4
langcodes==3.5.0
language_data==1.4.0
marisa-trie==1.3.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
msgpack==1.1.2
multidict==6.7.0
numpy==2.2.3
oauthlib==3.3.1
openpyxl==3.1.5
packaging==25.0
pandas==2.2.3
Pillow==9.1.1
prompt_toolkit==3.0.52
propcache==0.4.1
proto-plus==1.26.1
protobuf==6.33.1
psycopg2-binary==2.9.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.4
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.10.1
pyloudnorm==0.1.1
pyOpenSSL==25.1.0
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-monkey-business==1.1.0
python3-openid==3.2.0
pytz==2022.1
PyYAML==6.0.3
redis==7.0.1
referencing==0.37.0
requests==2.32.3
requests-oauthlib==2.0.0
rich==14.2.0
rpds-py==0.28.0
rsa==4.9.1
s3transfer==0.11.4
scipy==1.15.3
service-identity==24.2.0
shellingham==1.5.4
six==1.17.0
smart_open==7.4.4
sniffio==1.3.1
soundfile==0.13.1
sqlparse==0.5.3
srsly==2.5.2
tablib==3.8.0
tomli==2.3.0
tqdm==4.67.1
Twisted==24.11.0
txaio==23.1.1
typer==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
ua-parser==1.0.1
ua-parser-builtins==0.18.0.post1
uritemplate==4.2.0
urllib3==2.5.0
user-agents==2.2.0
uvicorn==0.17.6
uvloop==0.21.0
vine==5.1.0
wcwidth==0.2.14
whitenoise==6.7.0
wrapt==2.0.0
xlrd==2.0.1
XlsxWriter==3.2.2
xlwt==1.3.0
yarl==1.22.0
zope.interface==8.0.1
